orgnize the views and serilizers on the dashboard folder
