+++ futurex_openedx_extensions/dashboard/views/system.py
@@ -1,0 +1,340 @@
+"""System views for the dashboard app"""
+from __future__ import annotations
+
+from typing import Any, Dict
+from urllib.parse import parse_qs, urlencode, urlsplit, urlunsplit
+
+from common.djangoapps.student.models import get_user_by_username_or_email
+from django.core.exceptions import ObjectDoesNotExist, ValidationError
+from django.core.paginator import EmptyPage
+from django.db.models.query import QuerySet
+from django.http import JsonResponse
+from django.shortcuts import get_object_or_404, redirect, render
+from django_filters.rest_framework import DjangoFilterBackend
+from rest_framework import status as http_status
+from rest_framework import viewsets
+from rest_framework.response import Response
+from rest_framework.views import APIView
+
+from futurex_openedx_extensions.dashboard import serializers
+from futurex_openedx_extensions.dashboard.docs_utils import docs
+from futurex_openedx_extensions.helpers import clickhouse_operations as ch
+from futurex_openedx_extensions.helpers.constants import (
+    CLICKHOUSE_FX_BUILTIN_CA_USERS_OF_TENANTS,
+    CLICKHOUSE_FX_BUILTIN_ORG_IN_TENANTS,
+    COURSE_ACCESS_ROLES_SUPPORTED_READ,
+    FX_VIEW_DEFAULT_AUTH_CLASSES,
+)
+from futurex_openedx_extensions.helpers.converters import error_details_to_dictionary
+from futurex_openedx_extensions.helpers.filters import DefaultOrderingFilter, DefaultSearchFilter
+from futurex_openedx_extensions.helpers.models import ClickhouseQuery, DataExportTask
+from futurex_openedx_extensions.helpers.pagination import DefaultPagination
+from futurex_openedx_extensions.helpers.permissions import (
+    FXHasTenantCourseAccess,
+    IsAnonymousOrSystemStaff,
+    IsSystemStaff,
+)
+from futurex_openedx_extensions.helpers.roles import (
+    FXViewRoleInfoMixin,
+    get_accessible_tenant_ids,
+    get_usernames_with_access_roles,
+)
+from futurex_openedx_extensions.helpers.tenants import (
+    get_all_tenants_info,
+    get_excluded_tenant_ids,
+    get_tenants_info,
+)
+
+default_auth_classes = FX_VIEW_DEFAULT_AUTH_CLASSES.copy()
+
+
+@docs('VersionInfoView.get')
+class VersionInfoView(APIView):
+    """View to get the version information"""
+    permission_classes = [IsSystemStaff]
+
+    def get(self, request: Any, *args: Any, **kwargs: Any) -> JsonResponse:  # pylint: disable=no-self-use
+        """
+        GET /api/fx/version/v1/info/
+        """
+        import futurex_openedx_extensions  # pylint: disable=import-outside-toplevel
+        return JsonResponse({
+            'version': futurex_openedx_extensions.__version__,
+        })
+
+
+@docs('AccessibleTenantsInfoView.get')
+class AccessibleTenantsInfoView(APIView):
+    """View to get the list of accessible tenants"""
+    permission_classes = [IsAnonymousOrSystemStaff]
+
+    def get(self, request: Any, *args: Any, **kwargs: Any) -> JsonResponse:  # pylint: disable=no-self-use
+        """
+        GET /api/fx/accessible/v1/info/?username_or_email=<usernameOrEmail>
+        """
+        username_or_email = request.query_params.get('username_or_email')
+        try:
+            user = get_user_by_username_or_email(username_or_email)
+        except ObjectDoesNotExist:
+            user = None
+
+        if not user:
+            return JsonResponse({})
+
+        tenant_ids = get_accessible_tenant_ids(user)
+        return JsonResponse(get_tenants_info(tenant_ids))
+
+
+@docs('AccessibleTenantsInfoViewV2.get')
+class AccessibleTenantsInfoViewV2(FXViewRoleInfoMixin, APIView):
+    """View to get the list of accessible tenants version 2"""
+    authentication_classes = default_auth_classes
+    permission_classes = [FXHasTenantCourseAccess]
+    fx_view_name = 'accessible_info'
+    fx_view_description = 'api/fx/accessible/v2/info/: Get accessible tenants'
+
+    def get(self, request: Any, *args: Any, **kwargs: Any) -> JsonResponse:  # pylint: disable=no-self-use
+        """
+        GET /api/fx/accessible/v1/info/?username_or_email=<usernameOrEmail>
+        """
+        username_or_email = request.query_params.get('username_or_email')
+        try:
+            user = get_user_by_username_or_email(username_or_email)
+        except ObjectDoesNotExist:
+            user = None
+
+        if not user:
+            return JsonResponse({})
+
+        tenant_ids = get_accessible_tenant_ids(user)
+        return JsonResponse(get_tenants_info(tenant_ids))
+
+
+@docs('DataExportManagementView.list')
+@docs('DataExportManagementView.partial_update')
+@docs('DataExportManagementView.retrieve')
+class DataExportManagementView(FXViewRoleInfoMixin, viewsets.ModelViewSet):  # pylint: disable=too-many-ancestors
+    """View to list and retrieve data export tasks."""
+    authentication_classes = default_auth_classes
+    permission_classes = [FXHasTenantCourseAccess]
+    serializer_class = serializers.DataExportTaskSerializer
+    pagination_class = DefaultPagination
+    fx_view_name = 'exported_files_data'
+    fx_default_read_only_roles = ['staff', 'instructor', 'data_researcher', 'org_course_creator_group']
+    fx_default_read_write_roles = ['staff', 'instructor', 'data_researcher', 'org_course_creator_group']
+    fx_allowed_write_methods = ['PATCH']
+    fx_view_description = 'api/fx/export/v1/tasks/: Data Export Task Management APIs.'
+    http_method_names = ['get', 'patch']
+    filter_backends = [DjangoFilterBackend, DefaultOrderingFilter, DefaultSearchFilter]
+    filterset_fields = ['related_id', 'view_name']
+    ordering = ['-id']
+    search_fields = ['filename', 'notes']
+
+    def get_queryset(self) -> QuerySet:
+        """Get the list of user tasks."""
+        return DataExportTask.objects.filter(
+            user=self.request.user,
+            tenant__id__in=self.fx_permission_info['view_allowed_tenant_ids_any_access']
+        )
+
+    def get_object(self) -> DataExportTask:
+        """Override to ensure that the user can only retrieve their own tasks."""
+        task_id = self.kwargs.get('pk')  # Use 'pk' for the default lookup
+        task = get_object_or_404(DataExportTask, id=task_id, user=self.request.user)
+        return task
+
+
+@docs('ExcludedTenantsView.get')
+class ExcludedTenantsView(APIView):
+    """View to get the list of excluded tenants"""
+    authentication_classes = default_auth_classes
+    permission_classes = [IsSystemStaff]
+
+    def get(self, request: Any, *args: Any, **kwargs: Any) -> JsonResponse:  # pylint: disable=no-self-use
+        """Get the list of excluded tenants"""
+        return JsonResponse(get_excluded_tenant_ids())
+
+
+@docs('TenantInfoView.get')
+class TenantInfoView(FXViewRoleInfoMixin, APIView):
+    """View to get the list of excluded tenants"""
+    authentication_classes = default_auth_classes
+    permission_classes = [FXHasTenantCourseAccess]
+    fx_view_name = 'tenant_info'
+    fx_default_read_only_roles = COURSE_ACCESS_ROLES_SUPPORTED_READ.copy()
+    fx_view_description = 'api/fx/tenants/v1/info/<tenant_id>/: tenant basic information'
+
+    def get(
+        self, request: Any, tenant_id: str, *args: Any, **kwargs: Any,
+    ) -> JsonResponse | Response:
+        """Get the tenant's information by tenant ID"""
+        if int(tenant_id) not in self.request.fx_permission_info['view_allowed_tenant_ids_any_access']:
+            return Response(
+                error_details_to_dictionary(reason='You do not have access to this tenant'),
+                status=http_status.HTTP_403_FORBIDDEN,
+            )
+
+        result = {'tenant_id': int(tenant_id)}
+        result.update(get_all_tenants_info()['info'].get(int(tenant_id)))
+        return JsonResponse(result)
+
+
+class ClickhouseQueryView(FXViewRoleInfoMixin, APIView):
+    """View to get the Clickhouse query"""
+    authentication_classes = default_auth_classes
+    permission_classes = [FXHasTenantCourseAccess]
+    fx_view_name = 'clickhouse_query_fetcher'
+    fx_default_read_only_roles = ['staff', 'instructor', 'data_researcher', 'org_course_creator_group']
+    fx_view_description = 'api/fx/query/v1/<scope>/<slug>: Get result of the related clickhouse query'
+
+    @staticmethod
+    def get_page_url_with_page(url: str, new_page_no: int | None) -> str | None:
+        """
+        Get the URL with the new page number
+
+        :param url: The URL
+        :type url: str
+        :param new_page_no: The new page number
+        :type new_page_no: int | None
+        :return: The URL with the new page number
+        :rtype: str | None
+        """
+        if new_page_no is None:
+            return None
+
+        url_parts = urlsplit(url)
+        query_params = parse_qs(url_parts.query)
+
+        page_size = query_params.get(DefaultPagination.page_size_query_param, None)
+        if page_size:
+            del query_params[DefaultPagination.page_size_query_param]
+
+        if 'page' in query_params:
+            del query_params['page']
+
+        if page_size:
+            query_params[DefaultPagination.page_size_query_param] = page_size
+        query_params['page'] = [str(new_page_no)]
+
+        new_query_string = urlencode(query_params, doseq=True)
+
+        new_url_parts = (url_parts.scheme, url_parts.netloc, url_parts.path, new_query_string, url_parts.fragment)
+        new_full_url = urlunsplit(new_url_parts)
+        return new_full_url
+
+    @staticmethod
+    def pop_out_page_params(params: Dict[str, str], paginated: bool) -> tuple[int | None, int]:
+        """
+        Pop out the page and page size parameters, and return them as integers in the result. Always return the page
+        as None if not paginated
+
+        :param params: The parameters
+        :type params: Dict[str, str]
+        :param paginated: Whether the query is paginated
+        :type paginated: bool
+        :return: The page and page size parameters
+        :rtype: tuple[int | None, int]
+        """
+        page_str: str | None = params.pop('page', None)
+        page_size_str: str = params.pop(
+            DefaultPagination.page_size_query_param, ''
+        ) or str(DefaultPagination.page_size)
+
+        if not paginated:
+            page = None
+        else:
+            page = int(page_str) if page_str is not None else page_str
+            page = 1 if page is None else page
+
+        return page, int(page_size_str)
+
+    def get(self, request: Any, scope: str, slug: str) -> JsonResponse | Response:
+        """
+        GET /api/fx/query/v1/<scope>/<slug>/
+
+        :param request: The request object
+        :type request: Request
+        :param scope: The scope of the query (course, tenant, user)
+        :type scope: str
+        :param slug: The slug of the query
+        :type slug: str
+        """
+        clickhouse_query = ClickhouseQuery.get_query_record(scope, 'v1', slug)
+        if not clickhouse_query:
+            return Response(
+                error_details_to_dictionary(reason=f'Query not found {scope}.v1.{slug}'),
+                status=http_status.HTTP_404_NOT_FOUND
+            )
+
+        if not clickhouse_query.enabled:
+            return Response(
+                error_details_to_dictionary(reason=f'Query is disabled {scope}.v1.{slug}'),
+                status=http_status.HTTP_400_BAD_REQUEST
+            )
+
+        params = request.query_params.dict()
+        self.get_page_url_with_page(request.build_absolute_uri(), 9)
+
+        page, page_size = self.pop_out_page_params(params, clickhouse_query.paginated)
+
+        orgs = request.fx_permission_info['view_allowed_any_access_orgs'].copy()
+        params[CLICKHOUSE_FX_BUILTIN_ORG_IN_TENANTS] = orgs
+        if CLICKHOUSE_FX_BUILTIN_CA_USERS_OF_TENANTS in clickhouse_query.query:
+            params[CLICKHOUSE_FX_BUILTIN_CA_USERS_OF_TENANTS] = get_usernames_with_access_roles(orgs)
+
+        error_response = None
+        try:
+            clickhouse_query.fix_param_types(params)
+
+            with ch.get_client() as clickhouse_client:
+                records_count, next_page, result = ch.execute_query(
+                    clickhouse_client,
+                    query=clickhouse_query.query,
+                    parameters=params,
+                    page=page,
+                    page_size=page_size,
+                )
+
+        except EmptyPage as exc:
+            error_response = Response(
+                error_details_to_dictionary(reason=str(exc)), status=http_status.HTTP_404_NOT_FOUND
+            )
+        except (ch.ClickhouseClientNotConfiguredError, ch.ClickhouseClientConnectionError) as exc:
+            error_response = Response(
+                error_details_to_dictionary(reason=str(exc)), status=http_status.HTTP_503_SERVICE_UNAVAILABLE
+            )
+        except (ch.ClickhouseBaseError, ValueError) as exc:
+            error_response = Response(
+                error_details_to_dictionary(reason=str(exc)), status=http_status.HTTP_400_BAD_REQUEST
+            )
+        except ValidationError as exc:
+            error_response = Response(
+                error_details_to_dictionary(reason=exc.message), status=http_status.HTTP_400_BAD_REQUEST
+            )
+
+        if error_response:
+            return error_response
+
+        if clickhouse_query.paginated:
+            return JsonResponse({
+                'count': records_count,
+                'next': self.get_page_url_with_page(request.build_absolute_uri(), next_page),
+                'previous': self.get_page_url_with_page(
+                    request.build_absolute_uri(),
+                    None if page == 1 else page - 1 if page else None,
+                ),
+                'results': ch.result_to_json(result),
+            })
+
+        return JsonResponse(ch.result_to_json(result), safe=False)
+
+
+class SetThemePreviewCookieView(APIView):
+    """View to set theme preview cookie"""
+    def get(self, request: Any) -> Any:  # pylint: disable=no-self-use
+        """Set theme preview cookie"""
+        next_url = request.GET.get('next', request.build_absolute_uri())
+        if request.COOKIES.get('theme-preview') == 'yes':
+            return redirect(next_url)
+
+        return render(request, template_name='set_theme_preview.html', context={'next_url': next_url})
